edit GetHealthyFoodUseCase and rename sum functions and class
