+++ src/main/kotlin/dependencies/AppModule.kt
@@ -3,9 +3,9 @@ package dependencies
 import data.csvHandler.MealsCsvParser
 import data.csvHandler.MealsCsvReader
 import data.csvHandler.Repository.MealsRepositoryImpl
-import logic.Repository.MealsFilter
+import logic.MealsFilter
 import logic.Repository.MealsRepository
-import logic.usecase.HealthyFastFoodFilter
+import logic.usecase.GetHealthyFoodUseCase
 import org.koin.dsl.module
 import java.io.File
 
@@ -23,10 +23,8 @@ val appModule = module {
     single <MealsRepository>{
         MealsRepositoryImpl(get(),get())
     }
-    single {
-        HealthyFastFoodFilter()
-    }
+
     single <MealsFilter> {
-        HealthyFastFoodFilter()
+        GetHealthyFoodUseCase(get())
     }
 }
