+++ src/main/kotlin/logic/MealsFilter.kt
@@ -1,0 +1,7 @@
+package logic
+
+import data.models.Meal
+
+interface MealsFilter {
+    fun getFilterMeals(): List<Meal>
+}
