--- src/main/kotlin/logic/Repository/MealsFilter.kt
@@ -1,7 +1,0 @@
-package logic.Repository
-
-import data.models.Meal
-
-interface MealsFilter {
-    fun getFilterMeals(meal: List<Meal>): List<Meal>
-}
